pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pylon"
version = "0.1.0"
description = "Multi-agent job hunting platform"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.0",
    "anthropic>=0.40",
    "structlog>=24.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.20",
    "fastapi>=0.115",
    "uvicorn>=0.32",
    "websockets>=14.0",
    "openpyxl>=3.1",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
gmail = ["google-api-python-client>=2.0", "google-auth-oauthlib>=1.0"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pylon"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100
